Not reserve the original key during rebuild without key_name
This patch will let the rebuild API not reserve the VM original
key pair during the rebuild API was called without key_name in
the API body. This means that if user does not explicitly input
the key_name, the rebuilt instance will not have a key pair.

Change-Id: I515e8928d41dbb3f027c7bbecb6a69e522c33abd
Jira-issue: OPENSTACK-526